[AWM] Id data type changed to int32_t

+++ bbque/app/working_mode.cc
@@ -42,7 +42,7 @@ WorkingMode::WorkingMode() :
 	strncpy(str_id, "", 12);
 }
 
-WorkingMode::WorkingMode(int8_t _id,
+WorkingMode::WorkingMode(int32_t _id,
 			 std::string const & _name,
 			 float _value,
 			 SchedPtr_t _owner) :
